feat: Group, Destination, Themes라우팅 및 파일 경로 수정/#54
